chore(ci): 接入 GitHub Actions CI/CD + clippy 清零 + 分层依赖违规修复(v0.3.4.0)
- CI/CD：新增 .github/workflows/ci.yml（源码行数→分层依赖→fmt→clippy 门禁 + 三平台 test/build）
  与 release.yml（v* tag 三平台打包并创建 GitHub Release）；新增 rust-toolchain.toml 锁定 Rust 1.96.1
- clippy 清零：cargo clippy --fix 自动修复 + 手动处理 dead_code（预留契约 API 以带注释的针对性
  allow 标注）与风格 lint，全仓库 253→0；ci.yml 的 Clippy 恢复 -D warnings 硬门禁
- 分层依赖修复：core/prediction/intra.rs 两处 backend::cpu 直调改为经 backend::ops::predict_plane 统一入口
- 格式：cargo fmt --all 全仓库统一
- 版本号 0.3.3.21→0.3.4.0（§13 四处同步：§13 表格 / crf-viewer+crf-cuda Cargo.toml / APP_VERSION / README）
- 文档：README CI/CD 状态更新、development.md 增 CI/CD 章节；补齐 LICENSE(MIT) 与 CONTRIBUTING.md
- 验证：cargo test --workspace 209 passed/0 failed；clippy -D warnings=0；fmt --check=0；两门禁脚本通过

+++ src-tauri/src/crf/core/illumination.rs
@@ -182,7 +182,10 @@ mod tests {
         assert_eq!(fit_predict(-100, 120, 0), -120);
         assert_eq!(fit_predict(-50, 80, 0), -40);
         // 饱和保护（i32 极值不溢出）
-        assert_eq!(fit_predict(i32::MAX, 120, 0), i32::MAX.saturating_mul(120) / 100);
+        assert_eq!(
+            fit_predict(i32::MAX, 120, 0),
+            i32::MAX.saturating_mul(120) / 100
+        );
     }
 
     #[test]
@@ -204,7 +207,7 @@ mod tests {
     #[test]
     fn test_search_identity_when_no_benefit() {
         // 纯差分无乘加收益：最优应接近 (100, 0)
-        let golden: Vec<i32> = (0..4096).map(|i| (i % 256) as i32).collect();
+        let golden: Vec<i32> = (0..4096).map(|i| i % 256).collect();
         // frame = golden + 少量随机扰动（无系统性乘加偏移）
         let frame: Vec<i32> = golden
             .iter()
@@ -221,7 +224,7 @@ mod tests {
     #[test]
     fn test_search_detects_multiplicative_shading() {
         // 构造光照渐变：frame = (0.92·golden) + 3（整帧乘加，与模型同形）
-        let golden: Vec<i32> = (0..8192).map(|i| (i % 200 + 20) as i32).collect();
+        let golden: Vec<i32> = (0..8192).map(|i| i % 200 + 20).collect();
         let frame: Vec<i32> = golden.iter().map(|&g| (g * 92) / 100 + 3).collect();
         let fit = search_lic(&golden, &frame).expect("scan must find a fit");
         assert!(fit.worthwhile(), "shading must be worthwhile: {fit:?}");
